MDEV-30689: JSON_SCHEMA_VALID for type=array return 1 for any string that starts with '['
Analysis: When type is non-scalar and the json document has syntax error
then it is not detected during validating type. And Since other validate
functions take const argument, the error state is not stored eventually.
Fix: there could be two ways of fixing it. Those validate() functions that
do not check whether the json document is syntactically correct can
be made to check it by scanning the json before returning the appropriate
result. But that could lead to repeated checking. Also if the
json is not syntactically correct in the first place then no need to check
for validation at all. So first only scan/go over the json doc and proceed
to validate only if there are no error.